requirements.txt
python-dotenv>=1.0.1
langchain>=0.2.10
langchain-openai>=0.1.16
openai>=1.35.10
chromadb>=0.5.5
tiktoken>=0.7.0
faiss-cpu>=1.8.0.post1; sys_platform != "darwin" or platform_machine != "arm64"
fastapi>=0.112.0
uvicorn>=0.30.5
streamlit>=1.36.0
pydantic>=2.8.2
